requirements.txt
alembic==1.11.1
amqp==5.1.1
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.1
billiard==3.6.4.0
blinker==1.8.2
botocore==1.29.142
cachelib==0.9.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
commonmark==0.9.1
coverage==7.5.1
cryptography==39.0.0
Deprecated==1.2.14
dnspython==1.16.0
eventlet==0.30.2
exceptiongroup==1.1.1
Flask==2.2.2
Flask-Mail==0.9.1
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
kombu==5.2.4
limits==3.12.0
Mako==1.2.4
MarkupSafe==2.1.1
mistune==2.0.5
nose==1.3.7
ordered-set==4.1.0
packaging==23.1
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
prompt-toolkit==3.0.36
pycparser==2.21
Pygments==2.15.1
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.3.1
pytest-cov==4.1.0
pytest-mock==3.11.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==4.3.4
python-http-client==3.3.7
pytz==2022.7
PyYAML==6.0
requests==2.28.1
rich==12.6.0
s3transfer==0.6.1
schedule==1.2.1
six==1.16.0
starkbank-ecdsa==2.2.0
tomli==2.0.1
typing_extensions==4.6.2
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.15.0
zipp==3.15.0